Update `/settings` endpoint (#722)
